requirements.txt
andrea_library~=0.1.0
python-dotenv~=1.0.1
requests~=2.32.3